Re-organised import structure, but more needs to be done in pycharm
