+++ assets/__init__.py
@@ -1,0 +1,3 @@
+from .enemy import Enemy
+from .enviroment import Enviroment
+from .room import Room
