+++ assets/enemy.py
@@ -1,0 +1,5 @@
+class Enemy:
+	def __init__(self, name, health):
+		self.name = name
+		self.health = health
+
