+++ assets/enviroment.py
@@ -1,0 +1,4 @@
+class Enviroment:
+	def __init__(self, name):
+		self.name = name
+
