+++ assets/room.py
@@ -1,0 +1,7 @@
+class Room:
+	def __init__(self, X, Y, Width, Height):
+		self.X = X
+		self.Y = Y
+		self.Width = Width
+		self.Height = Height
+
