+++ log.txt
@@ -1,2 +1,2 @@
-DEBUG:root:(52, 202)
+DEBUG:root:(63, 237)
 DEBUG:root:This message should go to the log file
